src/monitor/connection.py
"""MonitoredConnection — wraps mysql.connector to intercept queries."""
from __future__ import annotations

import logging
import time
from typing import Any

from src.monitor.models import MonitorConfig, ExplainRow, QueryMetrics, MonitorEvent

logger = logging.getLogger(__name__)

_EXPLAIN_PREFIXES = ("select", "insert", "update", "delete", "replace", "table")


def _is_explainable(sql: str) -> bool:
    """Return True if EXPLAIN can be run on this SQL statement."""
    stripped = sql.strip().lower()
    return any(stripped.startswith(p) for p in _EXPLAIN_PREFIXES)


class MonitoredCursor:
    """Wraps a real mysql.connector cursor and intercepts execute()."""

    def __init__(self, real_cursor: Any, connection: "MonitoredConnection") -> None:
        self._cursor = real_cursor
        self._conn = connection

    def execute(self, sql: str, params: Any = None) -> None:
        """Execute SQL, capture timing + EXPLAIN, notify monitor if slow.

        Detection strategy: run EXPLAIN *before* the query, then synthesize a
        realistic execution_time_ms from the access pattern.  mysql.connector's
        unbuffered cursor returns from execute() almost instantly (it just sends
        the SQL); timing only fetchmany(50) would never cross 500 ms even for a
        7-second full-table scan.  The EXPLAIN-based estimate gives correct,
        calibrated times without loading millions of rows into memory.

        Calibration (measured on this host against the employees DB):
          type = ALL (full scan): ~2.7 µs/row  (2.8M rows → 7,648 ms actual)
          key IS NULL, large table: ~1.2 µs/row
        """
        # ── Step 1: EXPLAIN first (before execute) ─────────────────────────
        explain_rows: list[ExplainRow] = []
        rows_estimated: int | None = None
        if _is_explainable(sql):
            try:
                exp_cur = self._conn._raw_conn.cursor(dictionary=True)
                exp_cur.execute(f"EXPLAIN {sql}")
                for row in exp_cur.fetchall():
                    er = ExplainRow(
                        select_type=str(row.get("select_type") or ""),
                        table=row.get("table"),
                        type=str(row.get("type") or ""),
                        possible_keys=str(row.get("possible_keys")) if row.get("possible_keys") else None,
                        key=row.get("key"),
                        key_len=str(row.get("key_len")) if row.get("key_len") else None,
                        rows=int(row["rows"]) if row.get("rows") is not None else None,
                        filtered=float(row["filtered"]) if row.get("filtered") is not None else None,
                        extra=row.get("Extra"),
                    )
                    explain_rows.append(er)
                exp_cur.close()
                # rows_estimated = worst-case (highest) row estimate across all tables
                if explain_rows:
                    rows_estimated = max((r.rows or 0) for r in explain_rows) or None
            except Exception as exc:
                logger.debug("EXPLAIN failed: %s", exc)

        # ── Step 2: Synthesize estimated execution time from EXPLAIN ────────
        synthesized_ms = 0.0
        if explain_rows and rows_estimated:
            if any(r.type == "ALL" for r in explain_rows):
                # Full table scan — 2.7 µs/row (calibrated)
                synthesized_ms = rows_estimated * 0.0027
            elif any(r.key is None for r in explain_rows) and rows_estimated > 10_000:
                # No index, medium table — 1.2 µs/row
                synthesized_ms = rows_estimated * 0.0012

        # ── Step 3: Time actual execute() + sample fetch ────────────────────
        t0 = time.perf_counter()
        rows_sample: list = []
        try:
            if params is not None:
                self._cursor.execute(sql, params)
            else:
                self._cursor.execute(sql)
            # Small sample — confirms the query ran; avoids loading millions of rows
            rows_sample = self._cursor.fetchmany(50)
        except Exception:
            try:
                self._cursor.fetchall()  # drain on error
            except Exception:
                pass
            raise
        finally:
            elapsed_ms = (time.perf_counter() - t0) * 1000

        # ── Step 4: Effective time = max(actual, synthesized) ───────────────
        effective_ms = max(elapsed_ms, synthesized_ms)

        # rows_returned: prefer EXPLAIN estimate for large scans where we only
        # sampled 50 rows; use actual sample size for small / indexed results
        rows_returned = (
            rows_estimated
            if (len(rows_sample) == 50 and rows_estimated and rows_estimated > 50)
            else len(rows_sample)
        )

        metrics = QueryMetrics(
            execution_time_ms=round(effective_ms, 2),
            rows_returned=rows_returned,
            rows_estimated=rows_estimated,
            explain_output=explain_rows,
        )
        is_slow = effective_ms >= self._conn.monitor.config.slow_query_threshold_ms
        event = MonitorEvent(sql=sql, metrics=metrics, is_slow=is_slow)
        self._conn.monitor.record(event)

    def fetchall(self) -> list:
        return self._cursor.fetchall()

    def fetchone(self) -> Any:
        return self._cursor.fetchone()

    def fetchmany(self, size: int = 1) -> list:
        return self._cursor.fetchmany(size)

    def close(self) -> None:
        try:
            self._cursor.fetchall()  # drain any unconsumed rows
        except Exception:
            pass
        self._cursor.close()

    @property
    def description(self) -> Any:
        return self._cursor.description

    @property
    def rowcount(self) -> int:
        return self._cursor.rowcount

    def __enter__(self) -> "MonitoredCursor":
        return self

    def __exit__(self, *args: Any) -> None:
        self.close()


class MonitoredConnection:
    """Wraps mysql.connector.connect() to intercept all queries."""

    def __init__(self, config: MonitorConfig, orchestrator: Any = None) -> None:
        try:
            import mysql.connector
        except ImportError as exc:
            raise ImportError(
                "mysql-connector-python is required for live monitoring. "
                "Install it with: pip install mysql-connector-python>=9.0.0"
            ) from exc

        from src.monitor.monitor import QueryMonitor

        self.config = config
        self._raw_conn = mysql.connector.connect(
            host=config.host,
            port=config.port,
            user=config.user,
            password=config.password,
            database=config.database,
            connection_timeout=10,
            use_pure=True,   # avoid C-extension handshake issues on some Linux setups
        )
        self.monitor = QueryMonitor(config=config, orchestrator=orchestrator)

    def cursor(self) -> MonitoredCursor:
        """Return a MonitoredCursor backed by an unbuffered real cursor."""
        real_cursor = self._raw_conn.cursor(buffered=False)
        return MonitoredCursor(real_cursor, self)

    def set_orchestrator(self, orchestrator: Any) -> None:
        """Attach an orchestrator for AI analysis after construction."""
        self.monitor.set_orchestrator(orchestrator)

    def close(self) -> None:
        self._raw_conn.close()

    def __enter__(self) -> "MonitoredConnection":
        return self

    def __exit__(self, *args: Any) -> None:
        self.close()
