requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
mcp>=1.0.0
openai>=2.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
numpy>=2.0.0
scipy>=1.17.0
rich>=13.0.0
click>=8.1.0
httpx>=0.28.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
textual>=1.0.0
mysql-connector-python>=9.0.0
